Create apl bang
At APL93, I put up a one-page poster with the title "APL! -- A simple,
intuitive, and terse rendition of APL as ASCII". Alas, it was successful
beyond my wildest dreams--it was stolen, TWICE. Not wanting to give in to this
act of terrorism, I made a hundred copies and handed them out. "APL bang" was
first posted to comp.lang.apl on March 30, 1992; an earlier but totally
different attempt called "APL slash bang" was posted on December 31, 1991.
Anyway, I'm prompted by the recent discussion on c.l.a. to post the contents
of "APL!" (slightly edited) for those who have not seen it.